Extracted make_celery out
